final changes for doing decoding and alignment of native recordings
Former-commit-id: 98720040b5b3b67d5a5b194391de1d64fb508117 [formerly 5f1774322c485e7dbb1c34b449834e65a6c8ee4f] [formerly e99fc6e98625869937a61009724987bdffc6d075]
Former-commit-id: f0d83b3b2a5addbe264fa424b537392e1b6120f7 [formerly 0867e19b4356b2e9999e4288d1265f68255871b7]
Former-commit-id: 959d8d9231bec3c5ea861e63f76647e78c7da6cd

+++ src/mitll/langtest/server/database/AnswerDAO.java
@@ -41,7 +41,7 @@ public class AnswerDAO extends DAO {
    *  @return id of new row in result table
    */
   public long addAnswer(Database database, int userID, String id, int questionID, String answer,
-                        String audioFile, boolean valid, String audioType, int durationInMillis,
+                        String audioFile, boolean valid, String audioType, long durationInMillis,
                         boolean correct, float pronScore, String deviceType, String device, String scoreJson, boolean withFlash) {
     Connection connection = database.getConnection(this.getClass().toString());
     try {
@@ -80,11 +80,11 @@ public class AnswerDAO extends DAO {
    * @param scoreJson
    * @param withFlash
    * @throws java.sql.SQLException
-   * @see #addAnswer(Database, int, String, int, String, String, boolean, String, int, boolean, float, String, String, String, boolean)
+   * @see #addAnswer(Database, int, String, int, String, String, boolean, String, long, boolean, float, String, String, String, boolean)
    */
   private long addAnswerToTable(Connection connection, int userid, String id, int questionID,
                                 String answer, String audioFile,
-                                boolean valid, String audioType, int durationInMillis,
+                                boolean valid, String audioType, long durationInMillis,
                                 boolean correct, float pronScore, String deviceType, String device, String scoreJson,
                                 boolean withFlash) throws SQLException {
   //  logger.debug("adding answer for exid #" + id + " correct " + correct + " score " + pronScore + " audio type " +audioType + " answer " + answer);
@@ -127,7 +127,7 @@ public class AnswerDAO extends DAO {
     statement.setBoolean(i++, true);
     statement.setBoolean(i++, true);
     statement.setString(i++, copyStringChar(audioType));
-    statement.setInt(i++, durationInMillis);
+    statement.setInt(i++, (int)durationInMillis);
 
     statement.setBoolean(i++, correct);
     statement.setFloat(i++, pronScore);
